Remove some duplicate fixtures
