+++ tests/support/pytest/helpers.py
@@ -175,6 +175,31 @@ def remove_stale_minion_key(master, minion_id):
 
 
 @pytest.helpers.register
+def remove_stale_master_key(master):
+    keys_path = os.path.join(master.config["pki_dir"], "master")
+    for key_name in ("master.pem", "master.pub"):
+        key_path = os.path.join(keys_path, key_name)
+        if os.path.exists(key_path):
+            os.unlink(key_path)
+        else:
+            log.debug(
+                "The master(id=%r) %s key was not found at %s",
+                master.id,
+                key_name,
+                key_path,
+            )
+    key_path = os.path.join(master.config["pki_dir"], "minion", "minion_master.pub")
+    if os.path.exists(key_path):
+        os.unlink(key_path)
+    else:
+        log.debug(
+            "The master(id=%r) minion_master.pub key was not found at %s",
+            master.id,
+            key_path,
+        )
+
+
+@pytest.helpers.register
 def remove_stale_proxy_minion_cache_file(proxy_minion, minion_id=None):
     cachefile = os.path.join(
         proxy_minion.config["cachedir"],
